fix: made custom hook, fixed patient.validate.ts
